Integrated boilerplate with db, global compose file
Seems like this was never actually tested, because it didn't work. This
commit integrates the boilerplate with the existing postgres database
code. Also created a global docker-compose which can bring up all of the
services at once. Still need to make sure the code is mounted as a bind
mount in dev image, and copied in prod image.