Made the app dir a bind mount
This is much more comfortable for development, as when you make small
changes to the source code, you don't have to rebuild the whole image
and reinstall all of the packages. It's not great for production, so we
should think of a way to separate the dev and production environments.