You no longer start this service without docker
Why? docker-compose creates a network for the containers, where each
container can contact another container using hostname only (e.g.
'postgres'). If you run it locally, without docker, you'll have to look
up the ip addresses for containers and change the code to include them,
which is a major pain in the ass.