Cart locking mechanism & modified update
The cart can now be protected via a locking mechanism, when you make a
request to the endpoint. It can be unlocked via another endpoint. The
status of the lock can be queried via a third endpoint. You cannot
modify a cart that is locked.
Furthermore, the cart modified datetime is now set accordingly.